% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conserveR-package.R
\docType{data}
\name{connectivity_network}
\alias{connectivity_network}
\title{connectivity_network
A bipartite network of methods included in the conserveR package linked by shared cited references.
Used for visualization in \code{\link{map_selection}}}
\format{
An object of class \code{network} of length 5.
}
\usage{
connectivity_network
}
\description{
connectivity_network
A bipartite network of methods included in the conserveR package linked by shared cited references.
Used for visualization in \code{\link{map_selection}}
}
\keyword{datasets}
