% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conserveR-package.R
\docType{data}
\name{edge}
\alias{edge}
\title{edge
An example dataset of the output of \code{\link{find_method}}. Conservation prioritization methods prioritized
by fit to an example set of data requirements.}
\format{
An object of class \code{data.frame} with 134 rows and 9 columns.
}
\usage{
edge
}
\description{
edge
An example dataset of the output of \code{\link{find_method}}. Conservation prioritization methods prioritized
by fit to an example set of data requirements.
}
\keyword{datasets}
