% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_method.R
\name{find_method}
\alias{find_method}
\title{Identify Most Suitable MEthods for Conservation Prioritization}
\usage{
find_method(
  ranking = c("both", "strict", "inclusive"),
  mca = TRUE,
  plot = TRUE
)
}
\arguments{
\item{ranking}{character string. The methods used for ranking the methods. See details. One of "both", "strict", "inclusive". Default = "both".}

\item{mca}{logical. If TRUE a multiple correspondence analysis is included, to plot the most suitable and related methods.}

\item{plot}{logical. IF TRUE the results of the mca are plotted.

Based on the \code{ranking} argument, the conservation prioritization methods in the database are
ranked according to the user-provided information. If ranking = "strict" methods receive
one point for each full agreement with user reply (yes and no), if ranking = "inclusive",
methods get one point when they include a feature confirmed by the user (but non for not including it).
This means that ranking = "inclusive" will  likely return more general methods
that can include many different types of data and perspectives. If ranking = "both",
methods are first ranked as in strict and then equal ranks split by the inclusive ranking.}
}
\description{
Identifies suitable methods for conservation prioritization based on a user dialogue on conservation targets and data availability.
}
\examples{
\dontrun{
find_method()
}

}
