% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conserveR-package.R
\docType{data}
\name{literature}
\alias{literature}
\title{Literature}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 134 rows and 43 columns.
}
\usage{
literature
}
\description{
Literature references for all conservation prioritization methods listed in \code{\link{traits}}, in bib2format.
}
\keyword{datasets}
