% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conserveR-package.R
\docType{data}
\name{mca_results}
\alias{mca_results}
\title{mca_results}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 134 rows and 7 columns.
}
\usage{
mca_results
}
\description{
Results of a multiple correspondence analysis of all methods included in the conserveR package based on the \code{\link{traits}} data set.
Used for visualization in \code{\link{map_selection}}.
}
\keyword{datasets}
